Fixed a problem with sorting of tags
Fixed a problem where sorting of multiple tags which where created on the same date would not work correctly. Now instead it sorts on the date first and if multiple tags are found on the same date it selects the tag with the "highest" version. For example, earlier if we had two tags with the names:
Application-v1.0.0
Application-v1.0.1
Which were created on the same date, the v1.0.0 tag would be selected. Now by also sorting on the refname it will select the tag with the highest version number, i.e Application-v1.0.1 instead.